feat: support importing compressed snapshot from url
