+++ utils/forest_utils/src/lib.rs
@@ -8,5 +8,6 @@ pub mod encoding;
 pub mod io;
 pub mod json;
 pub mod macros;
+pub mod miscs;
 pub mod monitoring;
 pub mod net;
